% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_df.R
\name{edit_df}
\alias{edit_df}
\title{Convert count data to data frame}
\usage{
edit_df(df)
}
\arguments{
\item{df}{}
}
\value{
data frame in long format
}
\description{
Convert count data to data frame
}
\examples{
LongCountDF <- edit_df(df)

}
