% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleCounts.R
\name{scaleCounts}
\alias{scaleCounts}
\title{ScaleCounts for RNAseq}
\usage{
scaleCounts(mat)
}
\arguments{
\item{mat}{}
}
\value{
Scaled matrix
}
\description{
ScaleCounts for RNAseq
}
\examples{
ScaledMat <- scaleCounts(mat)

}
